fix: data_preprocessing.py 코드 리팩토링

+++ data/data_preprocessing.py
@@ -1,6 +1,5 @@
 import pandas as pd
 from typing import Union
-from sklearn.experimental import enable_iterative_imputer
 from sklearn.impute import IterativeImputer, KNNImputer
 
 ### 이상치 탐지 함수 ###
@@ -14,11 +13,16 @@ def outlier_detection(data: pd.Series) -> pd.Series:
     Returns:
         pd.Series: 이상치에 해당하는 데이터 Series 반환
     """
+
+    # IQR 계산 
     Q1 = data.quantile(0.25)
-    Q3 = data.quantile(0.75)
+    Q3 = data.quantile(0.75) 
     IQR = Q3 - Q1
+
+    # 안 울타리 경계 지정 (기준 1.5 step)
     lower_bound = Q1 - 1.5 * IQR
     upper_bound = Q3 + 1.5 * IQR
+    
     return data[(data < lower_bound) | (data > upper_bound)]
 
 
@@ -41,6 +45,7 @@ def delete_low_density(
     Returns:
         pd.DataFrame: 선택할 구간에 포함하는 위치 중복도를 갖는 데이터를 선택 또는 삭제 후 데이터프레임 반환
     """
+
     # 위도, 경도 그룹별 개수 카운트
     group = data.groupby(["latitude", "longitude"])["index"].count()
     
@@ -125,7 +130,7 @@ class MissingValueImputer:
         # df가 DataFrame이 아닐 경우
         else:
             # 결측값을 중앙값으로 대체하고 반환
-           return df.fillna(df.median())    
+            return df.fillna(df.median())
 
 
     ### Iterative방법을 사용한 보간 함수 ###
@@ -192,7 +197,7 @@ class MissingValueImputer:
 
 
 ### 급처매물 제거 함수 ###
-def urgent_sale_cut(data: pd.DataFrame, sigma_weight: float) -> list[float]:
+def urgent_sale_cut(data: pd.DataFrame, sigma_weight: float) -> list[int]:
     """
     train_data에서 위도, 경도를 그룹화 하여 그룹별 가격에 대한 급처매물을 제거하는 함수
 
@@ -204,6 +209,7 @@ def urgent_sale_cut(data: pd.DataFrame, sigma_weight: float) -> list[float]:
         list[float]: train_data에서 급처매물인 index의 list 반환
     """
 
+    # 위도, 경도를 그룹별로 묶어 데이터프레임 생성
     grouped = data.groupby(["latitude", "longitude"])
     grouped_indices = grouped.groups
     value_list = list(grouped_indices.values())
